move test setup to it's own file
