+++ tests/test.js
@@ -1,46 +1,5 @@
 import { expect } from "chai";
-import { readdirSync, readFileSync } from "node:fs";
-import path from "node:path";
-import { JSDOM } from "jsdom";
-import { fileURLToPath } from "url";
-import { execa } from "execa";
-
-const __dirname = fileURLToPath(new URL(".", import.meta.url));
-
-const testApps = readdirSync(path.join(__dirname, "..", "test-apps"));
-
-async function buildApp(scenario) {
-  try {
-    return await execa("pnpm", ["build"], {
-      cwd: path.join(__dirname, "..", "test-apps", scenario),
-    });
-  } catch (error) {
-    throw new Error(
-      "Building the test app failed. Please test manually running `pnpm build` for that test app",
-      { cause: error },
-    );
-  }
-}
-
-const scenarios = testApps.filter((file) => {
-  // skip this app as it's intended to fail the build
-  return file !== "external-resource-missing-integrity-hash";
-});
-
-console.log("Building scenarios:");
-await Promise.all(
-  scenarios.map((scenario) =>
-    (async () => {
-      try {
-        await buildApp(scenario);
-        console.log(`✅ Built ${scenario}`);
-      } catch (error) {
-        console.error(`❌ Failed to build ${scenario}`);
-        throw error;
-      }
-    })(),
-  ),
-);
+import { scenarios, readFile, buildApp } from "./setup.js";
 
 describe("Embroider build", function () {
   let indexHtml;
@@ -48,20 +7,7 @@ describe("Embroider build", function () {
   for (const scenario of scenarios) {
     describe(`Scenario: ${scenario}`, function () {
       before(async function () {
-        try {
-          const indexHtmlContent = readFileSync(
-            path.join("..", "test-apps", scenario, "dist", "index.html"),
-            {
-              encoding: "utf8",
-            },
-          );
-          indexHtml = new JSDOM(indexHtmlContent).window.document;
-        } catch (error) {
-          throw new Error(
-            "Parsing index.html of test-app build failed. Double check that test-app has been built successfully.",
-            { cause: error },
-          );
-        }
+        indexHtml = await readFile(scenario);
       });
 
       it("adds integrity attribute to link elements for stylesheets", function () {
@@ -128,26 +74,9 @@ describe("External resource handling", function () {
     let indexHtml;
 
     before(async function () {
-      try {
-        const indexHtmlContent = readFileSync(
-          path.join(
-            "..",
-            "test-apps",
-            "external-resource-with-integrity-hash",
-            "dist",
-            "index.html",
-          ),
-          {
-            encoding: "utf8",
-          },
-        );
-        indexHtml = new JSDOM(indexHtmlContent).window.document;
-      } catch (error) {
-        throw new Error(
-          "Parsing index.html of test-app build failed. Double check that test-app has been built successfully.",
-          { cause: error },
-        );
-      }
+      indexHtml = indexHtml = await readFile(
+        "external-resource-with-integrity-hash",
+      );
     });
 
     it("Leaves existing crossorigin attribute as is", function () {
